Update version 0.0.7

+++ src/main/java/com/app/ua/repository/StudentRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * Author Vladislav Domaniewski 04
  */
 
-public interface StudentRepository extends JpaRepository <StudentEntity, Integer> {
+public interface StudentRepository extends JpaRepository <StudentEntity, Long> {
 }
